removing obsolete code, refactored string ext out
