+++ xc.lua
@@ -25,54 +25,6 @@ end
 -- Import needed modules.
 require("controller")
 
---- Extends Lua's builtin string library to add Python's .split method
----@param str any @the string to split
----@param delim any @the delimiter to split on
----@return table @the split elements
-function string.split(str, delim)
-    local out = {}
-    local pattern = string.format("(.-)%s(.*)", delim)
-    local i = 1
-
-    while str do
-        local part, remainder = string.match(str, pattern)
-        if part then
-            out[i] = part
-            str = remainder
-            i = i + 1
-        else
-            out[i] = str
-            break
-        end
-    end
-
-    return out
-end
-
---- OVERLOADED FUNCTION: Returns the first parameter stripped of all leading and trailing occurrences of the second (if present).
---- If no second parameter is passed, returns the first parameter stripped of leading and trailing spaces.
----@param str string @The string to strip
----@param ... string @The pattern to strip from string
----@return string @The string stripped of leading and trailing occurrences of pattern, if provided, stripped of leading and trailing spaces if not.
-function string.strip(str, ...)
-    local args = {...}
-    local paramCount = #args
-    local overloads = {}
-    overloads[1] = function(strip)
-        local stripped = strip:match("^%s*(.-)%s*$")
-        return stripped
-    end
-    overloads[2] = function(strip, pattern)
-        local stripped = strip:match(string.format("^%s*(.-)%s*$", pattern))
-        return stripped
-    end
-    if overloads[paramCount] == nil then
-        error(string.format("No overload defined for %d parameter signature of string.strip", #paramCount))
-    end
-    local func = overloads[paramCount]
-
-    return func(str, table.unpack(args))
-end
 
 -- Global Mach4 instance
 inst = mc.mcGetInstance()
